Improving the Enemy base
Implemented a Enemy Base init() which will modify some values based on the current level e.g. damage and health.

Improved the enemy collisions so their damage still has the same reach but they can group up more. I feel this is more visually appealing.

Added time and health UI and the game manager is in charge of the player health so the enemies can just call to it to damage the player. Will have to implement a safer singleton at some point.

Added some more safety checks and using bools to explain code which is nice.

TODO:
~~1. Change slash box collider to a trigger that that when attacking an enemy they will not move the player~~
2. Implement better movement if colliding with something
3. Improve Attack implementation and add inheritance like the enemy class currently uses to that it can be more variable
~~5. Damage for both enemies and the player, currently it is 1 hit kill for the enemies and there is no damage for the player~~
6. Add audio
7. Add more enemies
8. Add a shader effect for when a enemy dies or a simple particle effect?
9. Add damage to the player effect so you can tell when you take damage rather than the counter just going down
10. Improve enemy movement (boids)
11. rest of the game.